pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-auth-pre-cog"
version = "0.1.0"
description = "FHIR-native MCP server for AI-driven Prior Authorization reasoning"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "mcp>=1.0.0",
    "sqlalchemy>=2.0.30",
    "faker>=25.0.0",
    "fhir.resources>=7.1.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.35.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-mock>=3.14.0",
    "playwright>=1.44.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.package-dir]
"" = "."
